Extract TaskManagementExecutor from SqlTaskManager
Before this change, task management executor was a private property of
SqlTaskManager. To handle immediate memory revocation on MemoryPool
over-allocation, the executor needs to be exposed to
MemoryRevokingScheduler.
